Move relevant enums from types to constants
