+++ src/clients/Core/augments/SetExpenditurePayoutModifier.ts
@@ -6,7 +6,8 @@ import {
   IColonyV6,
   IColonyV7,
 } from '../../../contracts/IColony/exports';
-import { ColonyRole, TxOverrides } from '../../../types';
+import { ColonyRole } from '../../../constants';
+import { TxOverrides } from '../../../types';
 import { AugmentedIColony, getPermissionProofs } from './commonAugments';
 
 // Colonies that support this method
