+++ src/clients/Core/augments/SetPaymentDomain.ts
@@ -5,7 +5,8 @@ import {
   IColonyV2,
   IColonyV3,
 } from '../../../contracts/IColony/exports';
-import { ColonyRole, TxOverrides } from '../../../types';
+import { ColonyRole } from '../../../constants';
+import { TxOverrides } from '../../../types';
 import { AugmentedIColony, getPermissionProofs } from './commonAugments';
 
 // Colonies that support this method
