+++ src/clients/Core/augments/augmentsV3.ts
@@ -10,7 +10,8 @@ import {
   IColonyV9,
   IColonyV10,
 } from '../../../contracts/IColony/exports';
-import { ColonyRole, TxOverrides } from '../../../types';
+import { ColonyRole } from '../../../constants';
+import { TxOverrides } from '../../../types';
 import { ColonyNetworkClient } from '../../ColonyNetworkClient';
 import {
   addAugments as addCommonAugments,
