+++ src/clients/Core/augments/augmentsV4.ts
@@ -9,7 +9,8 @@ import {
   IColonyV9,
   IColonyV10,
 } from '../../../contracts/IColony/exports';
-import { ColonyRole, TxOverrides } from '../../../types';
+import { ColonyRole } from '../../../constants';
+import { TxOverrides } from '../../../types';
 import { ColonyNetworkClient } from '../../ColonyNetworkClient';
 import { getPermissionProofs, AugmentedIColony } from './commonAugments';
 import {
