+++ src/clients/Extensions/OneTxPayment/augments/commonAugments.ts
@@ -7,7 +7,8 @@ import {
 
 import { OneTxPaymentVersion } from '../exports';
 import { AnyOneTxPayment } from '../../../../contracts/OneTxPayment/exports';
-import { ClientType, ColonyRole, TxOverrides } from '../../../../types';
+import { ClientType, ColonyRole } from '../../../../constants';
+import { TxOverrides } from '../../../../types';
 import {
   OneTxPaymentEvents,
   OneTxPaymentEvents__factory as OneTxPaymentEventsFactory,
