+++ src/clients/Extensions/ReputationBootstrapper/augments/commonAugments.ts
@@ -2,7 +2,7 @@ import { AugmentedIColony } from '../../../Core/augments/commonAugments';
 
 import { ReputationBootstrapperVersion } from '../exports';
 import { AnyReputationBootstrapper } from '../../../../contracts/ReputationBootstrapper/exports';
-import { ClientType } from '../../../../types';
+import { ClientType } from '../../../../constants';
 import {
   ReputationBootstrapperEvents,
   ReputationBootstrapperEvents__factory as ReputationBootstrapperEventsFactory,
