+++ src/clients/Extensions/Whitelist/augments/commonAugments.ts
@@ -2,7 +2,7 @@ import { AugmentedIColony } from '../../../Core/augments/commonAugments';
 
 import { WhitelistVersion } from '../exports';
 import { AnyWhitelist } from '../../../../contracts/Whitelist/exports';
-import { ClientType } from '../../../../types';
+import { ClientType } from '../../../../constants';
 import {
   WhitelistEvents,
   WhitelistEvents__factory as WhitelistEventsFactory,
